Bluetooth: Allow reconnection from Host on paired HID Device.
Host can send connection request when paired HID device
is in disconnected state. we are not checking the
NORMALLY_Connectible flags from sdp record, and anyways
sending this request from host, for subsequent user initiated
connection. If the remote HID is not in pagescan mode, we will
2 retries to connect, before giving up. If the remote HID is
in page scan mode, it will get connected.

CRs-fixed: 463916, 478811

Change-Id: I2b3c5c25dc7b08ab6ed0c3667897e5ea0f05f914