Bluetooth: Avoid to advertise PANU service records.
This is special case where remote device(PC) acts as PANU
& is trying to connect PANU on DUT instead of NAP. Because
DUT had advertised PANU service records on SDP. As DUT does
not support PANU-PANU connection, So connection is dropped
every time. As per spec, It's not mandatory to advertise PANU
service records, So PANU service record should not be advertised
on SDP to solve this issue.

Change-Id: I1f733467335b5e0ca00899cf4380860d233aab33
CRs-Fixed: 520804