Bluetooth: Reduce transmit windows if ACL buffer is shared
If BLE doesn't have dedicated buffers in the controller, then we
need to share one of the buffers out of the BR/EDR Quota. In that
case, transmit window for the BR/EDR should be reduced accordingly.
Without this change, BTU task won't allow the controller to sleep
due to condition mismatch.

CRs-Fixed: 589335
Change-Id: Ica4a4a59a10c9c89cc53a5c37c344f41405d741c