Bluetooth: GAP: Update Paging and inquiry busy level sepeately
While updating the busy level of Paging or inquiry '10' was
used as BL between paging and inquiry,this leads to ignore
the one of them when both of them are acitve.

Made unique busy level for paging related events and inquiry
related events, so that each event will be notified.

CRs-Fixed: 580854,570163

Change-Id: I7bddcb4901061c450555235717b7d49bab922ae8