Bluetooth: Send service change indication to paired dev at reboot
This patch adds changes to add the paired devices to the device list
which will be sent notifications for service change once the devices
are connected. This is needed since at device reboot, new handles will
be allocated to the services which might not be same as the one's
cached by the remote device.

CRs-Fixed: 638546
Change-Id: I091b06de4e3d85ec0471e58b6a8187d4cabcac0f
(cherry picked from commit 447ddce1a796c52880db3f375dabd6b4e8737fc3)