Dequeues direct connection request during cancel conn operation
When cancel connection request for a device is sent from an app
and if the current pending connection request's BD address does
not match with this device, the entries of the connection request
pending queue are also checked. If BD address match occurs with an
entry in the connection request queue, the entry is removed from
the queue.

CRs-fixed: 715606
Change-Id: I1bf50a424d86ac53a5201fff742c822f4c8d1c0b