Do not initiate SDP for incoming AVDTP connection
Do not initiate SDP for incoming AVDTP connection.
There are few instances when remote initiates L2cap
channel for AVDTP signaling but keeps quite after that,
and later on timeout DUT goes for rest of the setup
procedure and does sdp prior to the same. Few peer
devices deletes sink sdp record when there is an
active connection / ongoing connection, as a result
sdp request fails and DUT does not proceed with connection.

This patch makes sure that DUT does not go for redundant
sdp request for AVDTP provided remote has already established
L2cap connection for the same.

Change-Id: I4ed5846b894258334c7b515ea62d2edffb4d55f3