Relax max number of A2dp packets to send to SOC
Relax number of A2dp packets to be sent to controller when
system is loaded and media thread could not be scheduled
on time to process PCM packets.
This patch will help in compensating for the loss incurred
because of thread delay to some extent.
Moreover for 2mbps case, based on iteration calculation,
in some iterations if DUT is required to send more than one
Avdtp packets then this change will handle the same effectively.

CRs-Fixed: 819381
Change-Id: Ida6e3fb226570cc7a4a09bbf02c61aa818a4f7d3