Bluetooth: Add change to have unique paired device entry w.r.t static addr
When remote device is paired , it has unique pid comprising of IRK and
static bd addres. When a factory reset is done on it, again new IRK is
generated. Now if it is paired again , we get new pid with new IRk and
static bd adress. So in paired device list we eventually see two entries
fo remote device with same static address. Add change to handle this case
and remove the older entry if at all found with same static address

Change-Id: I329f739f65aa894760a3381d9268b43bc459d2ed