Bluetooth: Update LE pairing state for peripheral
When LE pairing is initated, the pairing flag for the peripheral is
not set. The same is true for central, if the pairing is initiated
by the peripheral using security request. Although LE doesn't use
this flag, it can be used by BR/EDR to prevent another pairing
based on the BTM pairing state.

CRs-Fixed: 843508
Change-Id: Ib2dae982f28fff5a2c977b83309bb343c5fc39a8