Bluetooth: Maintain disconnect reason during pairing
This change fixes the issue where a failure to connect
occurs while trying to pair to the remote device.
Due to this the disconnection reason is maintained as L2CAP_CONN_CANCEL
(0x0100), which jumps over the UINT8 causing the reason to be
reset to 0 and this giving an incorrect callback with status as
successful pairing.

CRs-Fixed: 843472
Change-Id: I71ffc456725f1c547004e37eb37e400dd8575e64