Do not reset SSR parameters on SCO link.
When Headset is connected during active call, HFP is connected
first. Due to active call SCO is also established, when SCO is
established we disable SSR. When a2dp is connected after SCO
connection it enables SSR. This enabling of SSR is causing ACL
disconnect.

When there is new profile connection on same device which has SCO
we should not enable SSR. It should be enabled only when there
is no SCO.

Change-Id: I25263a148a13136174a67867a822ba3497a11f8e