Stop SCO close timer when held/waiting call becomes active
For AT+CHLD=1, SCO is not disconnected by AG when active call is
ended. Same SCO connection is used when held/waiting call becomes
active. We receive CIEV1,0 and CIEV1,1 in that order. SCO close timer
will start as soon as CIEV1,0 is received. We should stop it after
receiving CIEV1,1.

CRs-fixed: 893120
Change-Id: I71cecaca6fbdbb70d0b4a42343b9fa26a5430b75