HID: Prevent incoming and outgoing SDP collision
This patch handles collision cases of outgoing and incoming SDP
for HID Host. SDP is performed 2 times as per BD design, one after
successful authentication, and one in hid stack layer as sdp data
is not stored currently in stack. When there is outgoing and
incoming connections, SDP was not getting completed within SDP
timeout for HID layer, which was resulting in SDP failure for hid
host resulting in improper disconnection of hid interrupt channel.

Change-Id: I5b4813dc916696e2c0f76be27a85c1a89c2d8be0
CRs-Fixed: 683139